lab4/requirements.txt
azure-eventhub>=5.11.2 